Initial Github Release.
Includes an incomplete TChannel implementation in Java. So far, the
basic `Message` types have been defined along with most of their
Encode/Decode logic ( Codecs ). These `Message`s represent message
primitives that the TChannel protocol speaks.

A basic `MessageMultiplexer` is also nearly complete, and is responsible
for re-constructing large Arg{1,2,3} payloads, or streaming payloads
that are necessarily chunked across multiple message containers.

The API design ( client / server APIs ) is still very much a WIP, but
the basic ideas are outlined in the README.md, and will be incorporated
into the Server and Client examples as more of the API is completed. The
Server and Client examples will also serve as a reference implementation
for those wishing to use tchannel4j in their projects.