Nearly done with CallRequest{Continue} Codec.
Adds a lot of code to the CallRequesCodec; I'm thinking that once this
is done we'll need to refactor just to reduce how big this file is. I
also haven't started on the CallResponse{Continue} Codec but it's going
to be 99.9% the same, so we won't have to re-implement the fragmentation
logic, but we will still have to make this Codec a little more generic.

- Remove arg1 parsing from CallRequestContinue messages; because arg1
  CANNOT be fragmented it can never appear in a CallRequestContinue
message. The same change will apply to CallResposneContinue messages.
- Add some preliminary test. Following commits will include more tests.