Move from ABCs to Interfaces for Messages.
Remove the complex ABC chain for messages, and instead group similar
messages by Interface. This means we have some duplicate code e.g. all
messages must define their own `id` field, but the Object structure is
now fully defined per class and thus much easier to understand.

We're still not done with this diff, which serves as the meat of the
protocol, but we're doing a lot of busy work now that will pay off big
in the end.