Move all stateful message processing to MessageMultiplexer.
Previously we were attempting to decode full CallRequest/Response
objects in the decoder. This meant that the decoder had to keep track of
Message defragmentation states in order to read the correct args. To
avoid this, Call{Request,Response}{Coneinue} Messages a.k.a.
CallMessages now don't make any assumptions about their arg{1,2,3}
payload, and intead treat it as an opaque ByteBuf. This allows us to
lazily encode/decode CallMessages -> FullMessages and allow the
MessageMultiplexer to handle frag/defrag state. It also allows the Codec
to assume that whatever payload is passed to it from the
MessageMultiplexer is *already fragmented properly* e.g. it can just
blindly write its payload.

Obviously, this means that the MessageMultiplexer has taken on the
responsibility of both message frag/defrag and arg aggregation ( which are
*very* similar but differnt concerns ).