Add a CallResponse Codec. Detangle Codecs from similar types.
Adds a simple CallResponse codec. It's nearly identical to the
CallRequest codec, seeing as they are nearly the same message. Also
separate out Codecs that serve similar messages. This could be
considered anti-DRY, but it allows for simpler logic and testing ( and
should be a little less confusing for onboarding developers ). This also
means that there is now a 1-to-1 mapping between Messages and their
associated Codec.

- Move `checksum` encoding/decoding to CodecUtils to centralize the
  conditional ByteBuffer read/write ( the conditional read/write is dependent on the
ChecksumType )
- Mark all Codecs as `final` classes.