Move from Abstract FullMessage classes to Interfaces.
To decouple request/response in the API from the FullMessage types in
the MessageMultiplexer, we've made FullMessage an interface that expects
ByteBufs, and we've genericized Request and Response s.t. clients can
use the interface however they see fit.

The base Request/Response objects are concrete Raw{Request,Response}
objects that are parameterized with `ByteBuf`. This means that they
automatically conform to FullMessage so we can link the API and
MessageMultiplexer layer without using the same types.

Unfortunately, there isn't Golang-style interface inference so we do
actually need to make RawRequest and RawResponse implement FullMessage.