Cleanup InitResponse.
We somehow weren't handling InitResponse messages anywhere. Whoops. This
makes sure that the InitRequestInitator ( from the client ) doesn't
close the channel until it receives an InitResponse.

This also adds basic tests for the InitRequestInitiator. A better
protocol-level integration test is probably needed to make sure that
they can talk to each other properly.