Expose the Resposne Code param to the Response object.
CallResponse Messages carry a 'response' code payload, which we convert
into an Enum, either 'OK' or 'Error'. This diff exposes that param to
the Response objects, plubming it all the way through to the client. The
semantics of arg{1,2,3} messages is up to the client.

- Also refactor the Request/Response builders to require required fields
  in the constructor of the builder to ensure that they are always set.
Previously we'd throw an error on build which is only discoverable once
somebody attempts to run the builder. That is not cool. This cleans up
the API significantly.