Add required fields and plugins to deploy to Sonatype
Since one cannot simply upload a project to Maven Central, we need to go
through a 3rd party 'host'. There are a few available but it appears as
though both Netty and Google Guava use Sonatype so I went with them.
There are a few requirements, and the full guide lives here:
http://central.sonatype.org/pages/ossrh-guide.html.

- Update the artifactId to `tchannel-java`
- Move versiont to SemVer
- Update docs everywhere
- Change tchannel-benchmark name to `tchannel-benchmark` from the
  generic name from the JMH template