Fixing Javadoc build failures and checkstyle inconsistencies.
We're preparing to release v0.1.2, our first release to the central
repositories, which means we're going through a lot of new build phases
e.g. javadocs, gpg signing, maven releaseing, etc. I'm learning a lot
and hopefully we'll nail-down a quasi-sane workflow. But until then,
it's a lot of trial-and-error followed by frantic commits.