Hide lock of Connection
Connection is meant to be publicly accessed.
It can also be subclassed. Synchronisation lock should
be hidden. If someone synchronises on Connection we will
get into deadlock.

This slightly changes the semantics of the class as it
makes impossible for classes that are not inheriting from
this to lock and be notified by it. However it is highly
unlikely that such behaviour is right.